+++ examples/react/src/index.ts
@@ -1,4 +1,4 @@
-import HelloWorld from './components/HelloWrold'
+import HelloWorld from './components/HelloWorld'
 
 export { HelloWorld }
 export { default as App } from './App'
