+++ examples/react/src/main.ts
@@ -1,1 +1,1 @@
-export { default as HelloWorld } from './components/HelloWrold'
+export { default as HelloWorld } from './components/HelloWorld'
