--- examples/vue/src/global.d.ts
@@ -1,1 +1,0 @@
-declare type GlobalType = string | number
