+++ examples/vue/src/index.ts
@@ -40,5 +40,3 @@ export {
 }
 
 export default DefaultImport
-
-// export const dtsTypeTest: GlobalType = 1
