+++ package.json
@@ -1,5 +1,6 @@
 {
-  "name": "dts-plugin-mono",
+  "name": "vite-plugin-dts",
+  "version": "3.0.0",
   "type": "module",
   "license": "MIT",
   "author": "qmhc",
@@ -22,6 +23,19 @@
     "test:ts": "pnpm -C examples/ts build",
     "test:vue": "pnpm -C examples/vue build"
   },
+  "main": "dist/index.cjs",
+  "module": "dist/index.mjs",
+  "types": "dist/index.d.ts",
+  "exports": {
+    ".": {
+      "types": "./dist/index.d.ts",
+      "require": "./dist/index.cjs",
+      "import": "./dist/index.mjs"
+    }
+  },
+  "files": [
+    "dist"
+  ],
   "engines": {
     "node": "^14.18.0 || >=16.0.0"
   },
@@ -31,6 +45,24 @@
   },
   "bugs": {
     "url": "https://github.com/qmhc/vite-plugin-dts/issues"
+  },
+  "keywords": [
+    "vite",
+    "vite-plugin",
+    "ts",
+    "dts",
+    "typescript"
+  ],
+  "dependencies": {
+    "@microsoft/api-extractor": "^7.34.4",
+    "@rollup/pluginutils": "^5.0.2",
+    "@rushstack/node-core-library": "^3.55.2",
+    "@vue/language-core": "^1.8.1",
+    "debug": "^4.3.4",
+    "fast-glob": "^3.2.12",
+    "kolorist": "^1.7.0",
+    "unplugin": "^1.3.1",
+    "vue-tsc": "^1.8.1"
   },
   "devDependencies": {
     "@babel/types": "^7.21.4",
