--- packages/unplugin/package.json
@@ -1,36 +1,0 @@
-{
-  "name": "unplugin-dts",
-  "version": "0.0.0",
-  "type": "module",
-  "license": "MIT",
-  "author": "qmhc",
-  "scripts": {},
-  "exports": {
-    ".": {
-      "types": "./dist/index.d.ts",
-      "require": "./dist/index.cjs",
-      "import": "./dist/index.mjs"
-    }
-  },
-  "files": [
-    "dist"
-  ],
-  "keywords": [
-    "vite",
-    "vite-plugin",
-    "ts",
-    "dts",
-    "typescript"
-  ],
-  "dependencies": {
-    "@microsoft/api-extractor": "^7.34.4",
-    "@rollup/pluginutils": "^5.0.2",
-    "@rushstack/node-core-library": "^3.55.2",
-    "@vue/language-core": "^1.8.1",
-    "debug": "^4.3.4",
-    "fast-glob": "^3.2.12",
-    "kolorist": "^1.7.0",
-    "unplugin": "^1.3.1",
-    "vue-tsc": "^1.8.1"
-  }
-}
