--- packages/unplugin/src/rspack.ts
@@ -1,3 +1,0 @@
-import unplugin from '.'
-
-export default unplugin.rspack
