--- packages/unplugin/src/webpack.ts
@@ -1,3 +1,0 @@
-import unplugin from '.'
-
-export default unplugin.webpack
