--- packages/vite-plugin/src/index.ts
@@ -1,5 +1,0 @@
-import { dtsPlugin } from './plugin'
-
-export default dtsPlugin
-
-export type { PluginOptions } from './types'
