+++ tsconfig.json
@@ -20,7 +20,7 @@
     "packages",
     "scripts",
     "tests"
-  ],
+, "src"  ],
   "exclude": [
     "node_modules"
   ]
